requirements.txt
PySide6==6.7.2
python-dotenv==1.0.1
httpx==0.27.2
chromadb==0.5.5
pydantic==2.8.2
PyPDF2==3.0.1
rapidfuzz==3.9.6
jinja2==3.1.4
python-docx==1.1.2
apscheduler==3.10.4
uvicorn==0.30.6
fastapi==0.114.0
